// src/common/utils/enums/message.enum.ts
export enum EmessageMapping {
    GET_OK = 'GET_OK',
    CANCEL_OK = 'CANCEL_OK',
    CHANNEL_ERROR = 'CHANNEL_ERROR',
    SHOPPING_CART_ID_ERROR = 'SHOPPING_CART_ID_ERROR',
    ERROR_CONEXION_GET_SHOPPINGCART = 'ERROR_CONEXION_GET_SHOPPINGCART',
    ERROR_CONEXION_CANCEL_SHOPPINGCART = 'ERROR_CONEXION_CANCEL_SHOPPINGCART',
    SHOPPINGCART_NOTFOUND = 'SHOPPINGCART_NOTFOUND',    
    ERROR_SERVICE_DELETE = 'ERROR_SERVICE_DELETE',
    DEFAULT_ERROR = 'DEFAULT_ERROR',
    DEFAULT ='DEFAULT'
}